Graphql-server/graphql/requirements.txt
Flask
Flask-GraphQL
graphene-sqlalchemy
graphene
psycopg2-binary
SQLAlchemy


